requirements.txt
gradio>=3.44
python-dotenv>=1.0
requests>=2.31